server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
firebase-admin
requests
python-multipart

